ENH: Make the integer conversions in copy lengths explicit
Third of five commits clearing -Wsign-conversion.  22 hunks covering the
length argument of memcpy(), memset(), strncpy() and strncat() where the
length arrives as an int or int64_t.

    memcpy(NBL->bricks[idest], NBL->bricks[sindex[c - 1]], NBL->bsize);
    memcpy(*list, tmplist, (new_length - 1) * sizeof(nifti1_extension));
    memcpy(ext->edata, data, len);
    strncpy(dest, data, field->len);

As with the allocation sizes, where a count is multiplied by a sizeof the
cast goes on the count so the multiply is done at 64-bit width rather
than after a possible 32-bit overflow.

field->len comes from the static field tables compiled into the tools,
not from the file being read; the extension lengths are validated by
nifti_check_extension() first; and NBL->bsize is the product fixed in the
first commit of this group.

+++ cifti/afni_xml.c
@@ -1138,7 +1138,7 @@ strip_whitespace(const char * str, int slen)
     *buf = '\0';
   else
   {
-    strncpy(buf, str + ifirst, len - ifirst - ilast);
+    strncpy(buf, str + ifirst, (size_t)(len - ifirst - ilast));
     buf[len - ifirst - ilast] = '\0';
   }
 
@@ -1210,7 +1210,7 @@ append_to_string(char ** ostr, int * olen, const char * istr, int ilen)
   }
 
   /* copy, starting at old nul char (if any), and terminate */
-  strncpy((*ostr) + *olen - 1, istr, ilen);
+  strncpy((*ostr) + *olen - 1, istr, (size_t)ilen);
   (*ostr)[newlen - 1] = '\0';
   *olen = newlen;
