+++ niftilib/nifti1_tool.c
@@ -2997,7 +2997,7 @@ modify_field(void * basep, field_s * field, const char * data)
         /* otherwise, we're good */
         {
           const int16_t sval = (int16_t)val;
-          memcpy((char *)basep + field->offset + fc * sizeof(sval), &sval, sizeof(sval));
+          memcpy((char *)basep + field->offset + (size_t)fc * sizeof(sval), &sval, (size_t)sizeof(sval));
         }
         if (g_debug > 1)
           fprintf(stderr, "+d setting posn %d of '%s' to %d\n", fc, field->name, val);
@@ -3017,7 +3017,7 @@ modify_field(void * basep, field_s * field, const char * data)
         }
         {
           const int32_t ival = (int32_t)val;
-          memcpy((char *)basep + field->offset + fc * sizeof(ival), &ival, sizeof(ival));
+          memcpy((char *)basep + field->offset + (size_t)fc * sizeof(ival), &ival, (size_t)sizeof(ival));
         }
         if (g_debug > 1)
           fprintf(stderr, "+d setting posn %d of '%s' to %d\n", fc, field->name, val);
@@ -3036,7 +3036,7 @@ modify_field(void * basep, field_s * field, const char * data)
           return 1;
         }
         /* otherwise, we're good */
-        memcpy((char *)basep + field->offset + fc * sizeof(fval), &fval, sizeof(fval));
+        memcpy((char *)basep + field->offset + (size_t)fc * sizeof(fval), &fval, (size_t)sizeof(fval));
         if (g_debug > 1)
           fprintf(stderr, "+d setting posn %d of '%s' to %f\n", fc, field->name, fval);
         posn += nchars;
@@ -3048,9 +3048,9 @@ modify_field(void * basep, field_s * field, const char * data)
     {
       char * dest = (char *)basep + field->offset;
       nchars = (int)dataLength;
-      strncpy(dest, data, field->len);
+      strncpy(dest, data, (size_t)(field->len));
       if (nchars < field->len) /* clear the rest */
-        memset(dest + nchars, '\0', field->len - nchars);
+        memset(dest + nchars, '\0', (size_t)(field->len - nchars));
     }
     break;
   }
